:bug: fix(config): border min-width/max-width を CGFloat 化（小数 width-breathing を有効化） (#273)
audit2 config-01。border の min-width/max-width は `.int`/`asInt`/`Int?` で
宣言されていたが、doc・schema・sibling の `width`(.cgDbl) は小数 0.5–30 を
謳っていた。sill の asInt は TOML double を coerce しないため `min-width = 0.5`
は decode 時に黙って捨てられ、width-breathing が起動しなかった。schema も
`"type":"integer"` + `"minimum":0.5`(到達不能)という内部矛盾を抱えていた。

トミー選択 = A（小数化）:
- spec の min-width/max-width を `.int`→`.cgDbl`（`.scalar(.number)`/asDouble）
- borderMinWidth/borderMaxWidth: Int? → CGFloat?
- effectiveBorderMin/MaxWidth: 不要な CGFloat($0) cast を除去
- config.schema.json 再生成（min/max-width が `"number"` に・差分は該当2行のみ。
  exclude rule の max-width は別キーで integer のまま正しい）

これにより config.toml/README が既に謳っていた「each 0.5–30」が型と一致
（audit2 config-02 = doc 矛盾も同根で自動解消）。

テスト: FacetConfigTests に (1) 直接代入の小数 in-range 存続 (2) **decode 経路**で
`min-width = 0.5` が捨てられず 0.5 になる回帰 pin を追加（CLT で swift test 不可・CI 検証）。

+++ Sources/FacetCore/FacetConfig.swift
@@ -155,9 +155,10 @@ public struct FacetConfig: Sendable {
     public var borderColorCycleMs: Int?
     /// Width breathing: when BOTH are set (max > min), the border width
     /// oscillates min↔max over `cycle-seconds` (any effect). Unset →
-    /// the fixed `width`. Raw; read `effectiveBorderMin/MaxWidth`.
-    public var borderMinWidth: Int?
-    public var borderMaxWidth: Int?
+    /// the fixed `width`. Fractional like the sibling `width` (e.g.
+    /// `0.5`). Raw; read `effectiveBorderMin/MaxWidth`.
+    public var borderMinWidth: CGFloat?
+    public var borderMaxWidth: CGFloat?
 
     /// Per-mac-desktop `[desktop.N]` workspace configs. Outer key is
     /// the mac desktop ordinal (Mission Control order, 1-based, user
@@ -417,10 +418,10 @@ public struct FacetConfig: Sendable {
     /// Width-breathing bounds, px (each clamped 0.5–30), or `nil` when
     /// unset. Breathing runs only when BOTH are set and max > min.
     public var effectiveBorderMinWidth: CGFloat? {
-        borderMinWidth.map { max(0.5, min(30, CGFloat($0))) }
+        borderMinWidth.map { max(0.5, min(30, $0)) }
     }
     public var effectiveBorderMaxWidth: CGFloat? {
-        borderMaxWidth.map { max(0.5, min(30, CGFloat($0))) }
+        borderMaxWidth.map { max(0.5, min(30, $0)) }
     }
 
     /// Named-enum config values that were written but didn't match any
